IMP: refactoring template script w/dual os, removing input file parsing
