--- .github/workflows/cron-job-templating/jinja-template.j2
@@ -1,20 +1,0 @@
-name: cron-{{ INPUT_EPOCH }}-core-macos-latest
-
-env:
-  os: {{ env.os }}
-  distro: {{ env.distro }}
-  epoch: {{ INPUT_EPOCH }}
-
-schedule_on:
-  schedule: {{ schedule_on.schedule.cron }}
-
-jobs:
-  install_env:
-    name: "install qiime2-core={{ INPUT_EPOCH }} (macos-latest)"
-    runs-on: macos-latest
-    outputs: {{ jobs.install_env.outputs }}
-    steps: {{ jobs.install_env.steps }}
-      
-  test_package: {{ jobs.test_package }}
-
-  phone_home: {{ jobs.phone_home }}
